add setkey support to commands using api keys

+++ commands/matches.js
@@ -1,17 +1,26 @@
 const axios = require('axios');
+const tokenData = require('../schemas/tokenschema');
 
 module.exports = {
     name: "matches",
     description: "Accepts your bwtinder matches",
-    execute(client, message, args, Discord) {
-        message.delete()
-        let apiKey = args[0];
+    async execute(client, message, args, Discord) {
+        let apiKey;
+        let user = await tokenData.findOne({ userID: message.author.id});
+
+        if (args[0]) {
+            message.delete();
+            apiKey = args[0];
+        } else if (user) apiKey = user.apiToken;
+
         if (!apiKey) {
             message.channel.send(`Error: No API Key Specified. To get this, go to: https://bit.ly/3t8vTc9`)
         } else {
             axios.post('https://bwtinder.com/api/profile', {
                 token: apiKey
             }).then((res) => {
+                if (!res.data.success) return message.reply('`The api request failed, this may be because of an incorrect api token.`');
+
                 let matches = res.data.matches;
                 if (matches.length === 0) {
                     const noMatchesEmbed = new Discord.MessageEmbed()
